resume by check filesize and asncore

+++ srget4.py
@@ -19,6 +19,7 @@ def make_request(req_type, what, details, ver="1.1"):
     req_line = "{verb} {w} HTTP/{v}".format(
         verb=req_type, w=what, v=ver
     )
+    print req_line
     details = [
         "{name}: {v}".format(name=n,v=v) for (n,v) in details.iteritems()
     ]
@@ -33,46 +34,111 @@ def parse_URL(url):
         path = "/"
     return (ip,port,path)
 def decide_if_HEAD_or_GET_request(meta_doc,filename):
-    print meta_doc
-    print filename
+    
     if os.path.exists(meta_doc) and os.path.exists(filename):
         return "HEAD"
     else:
         return "GET"
 def create_meta_DocName(filename,path):
     return filename+".rdoc"+".txt"
-def downloadFromStart_or_resumeDownload(type_req,filename):
+
+def metaDoc_sameAs_newHeader(my_request,rdl,ip,port):
+    same_content = True
+    clientSocket = socket.socket(socket.AF_INET,socket.SOCK_STREAM) 
+    clientSocket.connect((ip,port))
+    clientSocket.send(my_request)
+    head_recv = clientSocket.recv(1024)
+    head_string = head_recv
+    while head_recv:
+        head_recv = clientSocket.recv(1024)
+        head_string+= head_recv
+    clientSocket.close()
+    head_string = head_string.split("\r\n")
+    for i in head_string[1:]:
+        i = i.split(':')
+        if i[0] == "Content-Length":
+        
+            if i[1].strip() != rdl[0].strip():
+                same_content= False
+             
+        elif i[0] == "Last-Modified":
+            
+            if i[1].strip() != rdl[1].strip():
+                same_content = False
+             
+
+        elif i[0] == "ETag":
+            
+            if i[1].strip() != rdl[2].strip():
+                same_content =False
+
+    return same_content
+    
+    
+
+
+def make_request2(GET_or_HEAD,ip,port,path,Brange,DEFAULT_PORT =80):
+    if port == None:
+        port = DEFAULT_PORT
+
+    return ("{GET_or_HEAD} {path} HTTP/1.1\r\nHost: {ip}:{port}\r\nConnection: Close\r\n{Brange}Accept: text/html\r\n\r\n".format(GET_or_HEAD = GET_or_HEAD,path = path,Brange =Brange,ip = ip,port = str(port)))
+
+
+
+def downloadFromStart_or_resumeDownload(type_req,ip,path,port,filename,meta_doc):
     print "inside downloadFromStart_or_resumeDownload"
     print type_req
-
-
+    Brange = 0
+    my_request = make_request2('GET',ip,port,path,Brange)
     if type_req == "GET":
         print "Removing filename"
-        os.remove(filename)
-        pass
+        if os.path.exists(filename):
+            os.remove(filename)
+        return my_request,Brange
         
     if type_req == "HEAD":
-        pass
-        #metaDoc_compare_newHeader(my_request,ip,port,meta_doc,filename)
+        print "Opening meta_doc_r"
+        meta_doc_r = open(meta_doc,'r')
+        rdl = meta_doc_r.readlines()
+        print 'create HEAD request'
+        my_request = make_request('HEAD', path,
+            {'Host': ip,
+             'Connection': 'close'}
+        )
+
+        print "Entering metaDoc_compare_newHeader"
+        same_content = metaDoc_sameAs_newHeader(my_request,rdl,ip,port)
+        if same_content:
+            Brange = rdl[-1].strip()
+            Brange = os.path.getsize(filename)
+            byteRange = "Range: bytes={byteRange}-\r\n".format(byteRange = Brange)
+            my_request = make_request2('GET',ip,port,path,byteRange)
+            
+            print Brange
+
+        return my_request,Brange
+
+        
 
 def write_header_respone(data_recv,meta_doc_w):
     content_length = -1
     header,body = data_recv.split("\r\n\r\n")
+    print header
     NL = "\r\n"
     for i in header.split("\r\n")[1:]:
         i = i.split(":")
         field,atrribute = i[0],i[1]
         if field == "Content-Length":
             content_length = int(atrribute)
-            print atrribute
+            
             meta_doc_w.write(atrribute+NL)
                 
         elif field == "ETag":
-            print atrribute
+           
             meta_doc_w.write(atrribute+NL)
 
         elif field == "Last-Modified":
-            print atrribute
+            
             meta_doc_w.write(atrribute+NL)
     return body,content_length
 
@@ -94,7 +160,9 @@ class HTTPClient(asyncore.dispatcher):
         print "decide_if_HEAD_or_GET_request"
         self.type_request = decide_if_HEAD_or_GET_request(self.meta_doc,self.filename)
         print "downloadFromStart_or_resumeDownload"
-        downloadFromStart_or_resumeDownload(self.type_request,self.filename)
+        my_request,Brange = downloadFromStart_or_resumeDownload(self.type_request,self.ip,self.path,self.port,self.filename,self.meta_doc)
+        
+        
         self.full_string = ''
         # Create a TCP socket to host at the right port
         print "connecting"
@@ -108,22 +176,18 @@ class HTTPClient(asyncore.dispatcher):
         print "open filename doc w"
         self.filename_w = open(self.filename,'a')
        
-        self.len_recv_bytes =0
-        self.content_lengt = -1
-        self.total_bytes = -1
+        self.len_recv_bytes = int(Brange)
+        self.content_length = -1
         self.EveryByteDownloaded = False
+        print self.len_recv_bytes
        
         # Create recv buffer and send buffer
         (self.recvbuf, self.sendbuf) = ("", "")
-        
         # Make an initial request & deliver it
-        print "make request"
-        request = make_request('GET', self.path,
-            {'Host': self.ip,
-             'Connection': 'close'}
-        )
+        
         self.port = int(self.port)
-        self.write(request)
+        print my_request
+        self.write(my_request)
         
         
 
@@ -137,11 +201,11 @@ class HTTPClient(asyncore.dispatcher):
         
  
     def handle_close(self):
+        print "In handle_close"
         print "removing meta_doc"
+
         print "total_bytes: "+str(self.len_recv_bytes)
         os.remove(self.meta_doc)
-
-        
         self.close()
 
  
@@ -154,30 +218,26 @@ class HTTPClient(asyncore.dispatcher):
         self.sendbuf = self.sendbuf[bytes_sent:]
  
     def handle_read(self):
+        
         recv_bytes = self.recv(HTTPClient.RECV_CHUNK_SIZE)
         self.full_string += recv_bytes
-
         if not self.header_found and "\r\n\r\n" in self.full_string:
-            
             self.header_found = True
             print "write_header_respone"
             self.body, self.content_length = write_header_respone(self.full_string,self.meta_doc_w)
             self.filename_w.write(self.body)
-           
-            self.len_recv_bytes = len(self.body)
+            self.len_recv_bytes += len(self.body)
             self.meta_doc_w.write(str(self.len_recv_bytes)+"\r\n")
-
         elif self.header_found and not self.EveryByteDownloaded:
             self.filename_w.write(recv_bytes)
             self.len_recv_bytes += len(recv_bytes)
-            self.meta_doc_w.write(str(self.len_recv_bytes)+"\r\n")
-            if self.content_length - self.len_recv_bytes  <= 0:
-                self.EveryByteDownloaded = True
+            #self.meta_doc_w.write(str(self.len_recv_bytes)+"\r\n")
+            
 
 
 
  
 if __name__ == "__main__":
     
-    clients = [HTTPClient("http://10.27.8.20:8080/primeNumber.hamuel","ok.txt")]
+    clients = [HTTPClient("http://10.27.8.20:8080/bigfile.xyz","ok.txt")]
     asyncore.loop()
